Remove conan
This is done for two main reasons:

a) The current cmake-conan integration
(https://github.com/conan-io/cmake-conan) is only stable for old
versions of conan. Its support for conan 2 is only experimental. I am
developing on Gentoo, where conan 2 is already the standard and older
versions are deprecated.

b) I would like this project to evetually become a Gentoo package, and
it is not a good idea to mix conan with portage for dependency
management.